Update loading of Masonry items
Images were not loading fast enough because the ImagesLoaded plugin was
waiting for all images to load before applying a Masonry layout. This
update assigns new width and height and keeping the aspect ratio of
resized images. It also assigns a background color to the items based
on the color of the image.  This color shows before the images are
fully loaded.This update loads images faster without affecting the
Masonry layout.